Create dynamic queues for reply destination
Create dynamic queues when ReplyDestination parameter is not specified.
The dynamic queue is deleted upon cosuming the reply message.
This is according to the JMS Message ID pattern.

If the fixed queue should be used (Correlation ID Pattern) following
parameter should be set to true.
transport.jms.CachedResponseQueue

Please note that above changes are effective only when the ReplyDestination
parameter is not set.